requirements.txt
# Núcleo
torch>=1.13
torchvision>=0.14
pandas>=1.4
pillow>=9.0

# Utilidades
tqdm>=4.64

# (Opcional, si quieres usar config.yaml y cargarlo bien)
pyyaml>=6.0

# (Opcional, si usas Jupyter para pruebas)
#jupyter

matplotlib